Start rewriting cancellation token
